% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/connect.sql.server.R
\name{connect.sql.server}
\alias{connect.sql.server}
\title{A function that connects an R session to a running instance of
Microsoft SQL Server using the jTDS driver and RJDBC.}
\usage{
connect.sql.server(database, user, password, server.address, domain = NA)
}
\arguments{
\item{database}{The name of the database on the server that you want to connect to}

\item{user}{The user name that you want to use to authenticate against SQL Server}

\item{password}{The password for the user that will be authenticated against SQL Server}

\item{server.address}{The network address of the Windows instance that is running SQL Server}

\item{domain}{The domain to use in the case of NTLM logins, NA if using SQL Server login credentials}
}
\description{
A function that connects an R session to a running instance of
Microsoft SQL Server using the jTDS driver and RJDBC.
}
\examples{
connect.sql.server()
}
\keyword{RJDBC}
\keyword{SQL}
\keyword{Server}

